requirements.txt
aiohttp==3.8.5
keras==2.13.1
keras==2.13.1rc0
matplotlib==3.7.2
matplotlib==3.6.3
numpy==1.24.3
numpy==1.24.1
onnxruntime==1.15.1
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.4.0
Pillow==10.0.0
python-socketio==5.9.0
Requests==2.31.0
roboflow==1.1.1
scikit_learn==1.3.0
supervision==0.11.1
tensorflow_macos==2.13.0rc1
torch==2.0.1
ultralytics==8.0.154
